rename: 대시보드 페이지 컴포넌트들 재배치(#99)
