+++ src/pages/dashboard/sub-pages/Join.tsx
@@ -1,7 +1,6 @@
 import type { DropdownProps } from '@/components/ui/Dropdown/Dropdown.types'
-import Section from '../common/Section'
+import { Section, DROPDOWN_PERIOD_OPTIONS } from './common'
 import { useState } from 'react'
-import { DROPDOWN_PERIOD_OPTIONS } from '../common/constants'
 import VerticalBarChart from '@components/charts/VerticalBarChart/VerticalBarChart'
 
 export default function Join() {
