+++ src/pages/dashboard/sub-pages/Leave.tsx
@@ -1,6 +1,5 @@
 import { useState } from 'react'
-import Section from '../common/Section'
-import { DROPDOWN_PERIOD_OPTIONS } from '../common/constants'
+import { Section, DROPDOWN_PERIOD_OPTIONS } from './common'
 import type { DropdownProps } from '@components/ui/Dropdown/Dropdown.types'
 import VerticalBarChart from '@components/charts/VerticalBarChart/VerticalBarChart'
