+++ src/pages/dashboard/sub-pages/common/index.ts
@@ -1,0 +1,2 @@
+export * from './constants'
+export { default as Section } from './Section'
