+++ src/pages/dashboard/sub-pages/index.ts
@@ -1,3 +1,3 @@
-export { default as Join } from './Join/Join'
-export { default as Leave } from './Leave/Leave'
-export { default as Reason } from './Reason/Reason'
+export { default as Join } from './Join'
+export { default as Leave } from './Leave'
+export { default as Reason } from './Reason'
